refactor: extract program-start logic from Server into a Main class
